Mayor Refactor of code. Tons of bugfixes. Stable version archieved
Added:
- Custom editor for making Data-classes.
- Overide save functionality.

+++ Assets/DataManagement/Scripts/Data/DataParser.cs
@@ -1,22 +1,26 @@
-﻿using UnityEngine;
-#if UNITY_EDITOR
-using UnityEditor;
-#endif
-
-using System.IO;
-using System.Text;
+﻿using System.IO;
 using System.Security.Cryptography;
+using System.Text;
+using UnityEngine;
 
 namespace DataManagement
 {
+    /// <copyright file="DataParser.cs">
+    /// Copyright (c) 2019 All Rights Reserved
+    /// </copyright>
+    /// <author>Kevin Hummel</author>
+    /// <date>18/03/2019 21:41 PM </date>
+    /// <summary>
+    /// This class handles getting the data in Unity and saving it too disk.
+    /// </summary>
     public static class DataParser
     {
         public static string Encrypt(string p_input)
         {
             if (DataManager.Instance.encrypt)
             {
-                var t_inputbuffer = Encoding.Unicode.GetBytes(p_input);
-                var t_outputBuffer = DES.Create().CreateEncryptor(DataReferences.key, DataReferences.iv).TransformFinalBlock(t_inputbuffer, 0, t_inputbuffer.Length);
+                byte[] t_inputbuffer = Encoding.Unicode.GetBytes(p_input);
+                byte[] t_outputBuffer = DES.Create().CreateEncryptor(DataReferences.key, DataReferences.iv).TransformFinalBlock(t_inputbuffer, 0, t_inputbuffer.Length);
                 return System.Convert.ToBase64String(t_outputBuffer);
             }
             else return p_input;
@@ -24,23 +28,15 @@ namespace DataManagement
 
         public static ScriptableObject CreateAsset<T>(string p_name) where T : ScriptableObject
         {
-            var t_asset = ScriptableObject.CreateInstance<T>() as T;
-
-            #if UNITY_EDITOR
-            AssetDatabase.CreateAsset(t_asset, "Assets/Temp_Assets/" + p_name + ".asset");
-
-            AssetDatabase.SaveAssets();
-            AssetDatabase.Refresh();
-            EditorUtility.FocusProjectWindow();
-            Selection.activeObject = t_asset;
-            #endif
+            T t_asset = ScriptableObject.CreateInstance<T>() as T;
 
             return t_asset;
         }
 
         public static void SaveJSON(string p_name, string p_info)
         {
-            var t_path = Application.persistentDataPath + "/" + DataManager.Instance.DataReferences.ID + "/" + p_name + ".json";
+            if (DataManager.Instance == null) return;
+            string t_path = Application.persistentDataPath + "/" + DataManager.Instance.ID + "/" + SceneManager.Instance.DataReferences.ID + "/" + p_name + ".json";
             if (!File.Exists(t_path)) File.Delete(t_path);
 
             using (FileStream fs = new FileStream(t_path, FileMode.Create))
